+++ IssueTracker/IssueTracker/Presenter/Controller/LoginViewController.swift
@@ -8,16 +8,9 @@ class LoginViewController: UIViewController {
     @IBOutlet weak var githubLogInButton: UIButton!
     @IBOutlet weak var logInButton: UIButton!
     
-    private lazy var errorAlert: UIAlertController =  {
-        let alert = UIAlertController(title: "Error", message: "로그인이 실패하였습니다.", preferredStyle: .alert)
-        alert.addAction(UIAlertAction(title: "확인", style: .default, handler: .none))
-        return alert
-    }()
-    
     override func viewDidLoad() {
         super.viewDidLoad()
         setupButton()
-        setupObserver()
     }
 }
 
@@ -37,21 +30,12 @@ private extension LoginViewController {
     
     private func setupGithubLoginButton() {
         githubLogInButton.rx.tap
-            .subscribe(onNext: { _ in
+            .subscribe(onNext: { [weak self] _ in
                 LoginManager.loginPost(API.githubLogin)
+                guard let redirectionVC = self?.storyboard?.instantiateViewController(withIdentifier: ViewControllerID.redirection) else { return }
+                redirectionVC.modalPresentationStyle = .fullScreen
+                self?.present(redirectionVC, animated: true, completion: nil)
             }).disposed(by: rx.disposeBag)
-    }
-    
-    private func setupObserver() {
-        LoginObserver.validLoginTry.addObserver().bind { [weak self] object in
-            guard let _ = object as? LoginDTO else { return }
-            self?.moveToNextVC()
-        }.disposed(by: rx.disposeBag)
-        
-        LoginObserver.InvalidLoginTry.addObserver().bind { [weak self] _ in
-            guard let errorAlert = self?.errorAlert else { return }
-            self?.present(errorAlert, animated: true, completion: nil)
-        }.disposed(by: rx.disposeBag)
     }
     
     private func moveToNextVC() {
